+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### V0.8.0
 - Added simulated annealing optimization algorithm for different objectives (by [amarshah1999](https://github.com/amarshah1999))
 - Added support for opta/statsperform MA files
+- Removed the `TackleEvent` class; tackles remain available as `"tackle"` in the `databallpy_event` column of the event data
 
 ## Version 0.7
 ### V0.7.3 (14/04/2026)